Convert red link entry point from CSS to JSON
We decided to maintain red links as regular CSS classes in a HTML
post-processing step for better styling opportunities this brings. This change
converts the former experimental CSS entry point into a JSON entry point that
returns just the red links for a given revision. It is also experimental and
subject to change (see inline comments).