Don't assume that the body returned by preq is a String
I'm changing preq to return buffers by default. This is more efficient, as
those are allocated off-heap. JS strings above a certain size (80k IIRC) are
directly allocated in the old space of the JS heap, so are not collected very
quickly or efficiently. To make this work in restbase, we need to call
.toString() when comparing the body with strings.