T76883: Mod loading: Fix the module location to mods/
As of this commit, all 'file' modules need to reside in the mods/
directory, unless an absolute path is given.

This commit also allows such modules not to provide the path stanza iff
they reside in the mods/ directory directly (not in a sub-dir) and their
file name matches their module name.