Draft for new module_info export structure
James & I discussed the interaction between specs, modules, dependencies and
resources a bit more, and this is the (preliminary) result.

We realized that required resources are pretty much implementation-specific,
so moved them into the module spec. We also realized that the use case of
specifying the spec separately from the module might not be as important as we
initially thought, as we can always create a module as a wrapper. Since the
module will want to include a shared spec / testing module for the spec
anyway, it looks simpler to just export the spec from the module.

This brings up the question of the relationship between the module and spec
versions. In semver, the major & middle version digits signify breaking and
non-breaking API changes respectively, so it seems fitting that we use those
to reflect the major & middle bits of the spec version. This leaves the minor
version on both the spec and implementation as independent version parts.