Resource support v1
Modules can now export an array of request templates in module_info.resources,
which will be executed after the full routing tree has been built. Default
method is PUT.

This mechanism can be used to create tables, buckets and other dynamic
resources. It could also be used to assert the presence of other resources in
the future, although module dependencies are probably better implemented as a
`.dependencies` property outlining the actual modules and their expected paths.