Introduce a key_rev_value bucket
This bucket indexes by MediaWiki revision along with the tid. This allows
direct access to content by MW revision id, which speeds up the common case.
Grouping by revision id also lets us use the current tid for our renders,
which more truthfully reflects the current 'as of now' template & other
resource usage.

For revision deletion & suppression we'll still need to check some other index
structure, as it doesn't really make sense to store this separately on
each property. One option would be to maintain a list or bloom index of
revdeleted revision ids in memory, and then check with the revision table for
the actual restrictions. The other option would be to simply perform a query
against the revision table in parallel with the main request, although that
would likely give away some of the performance gains.