Simple gzip encoding support
For JSON & text/* we definitely want to support gzip Content-Encoding. This
patch hooks up a basic version. Sadly this halves the throughput for 55k HTML
responses to around 1100/s on my laptop. We will want to have a varnish layer
in front to store the gzipped version of hot items.