Fix: Detect when a bad revision has been fed to the MW API
The action module expects the query.pages field to be present in the
response, otherwise it throws a 500 error with the description 'Missing
query pages from the PHP action API response', thus making it impossible
for page_revisions module to throw a 404.

This commit fixes this issue by catching any errors thrown by the action
module and inspecting its status and the error description. In the
aforementioned case, a 404 is thrown, otherwise the original error is
propagated.