Deny access to the content of restricted revisions
If a specific revision is passed in, a second, parallel request is made
to the page_revisions module to determine whether the sought revision is
accessible. In case it is restricted, access in denied to the user. Note
that the content itself is fetched and stored regardless of access
restrictions.