Use bluebird internally & don't rely on the global Promise
In ES6 the way to deal with different Promise implementations is to require
the local implementation without overwriting the global Promise object. At
library interfaces an incoming ES6 promise can be converted to an extended
Promise by wrapping it in the local Promise's Promise.resolve().

JSHint does not like us re-defining the Promise object, so we are using 'P'
instead for the local (bluebird) definition.