Don't swallow & generate content for non-404 errors
The unconditional catch handler in parsoid.js triggered content generation on
*any* errors, not just 404s. This led to a large number of revisions to be
stored on some pages I was using for benchmarking.

It also covered up a bug in the handling of requests for 'latest' revisions.
The frontend code (and a test) was creating internal requests with 'latest' as
the revision, but the revision parsing code in key_rev_value did not consider
that a valid revision. Combined with the earlier bug this meant that each
request for an existing version was causing a new revision to be parsed &
saved.

This patch solves both problems by

- only generating content on 404, and re-raising any other errors, and

- dropping the `/latest` pseudo-revision in favor of a straight request for
  the title.