Don't try to destructively update revision entries
When updating revision metadata on revision deletion / suppressions we need to
use a newer timeuuid so that our revisioning works properly. If we try to
overwrite the entry with the same tid we'll end up using the same writetime as
well, which means that the save doesn't end up happening.

As a consequence of proper versioning we now need to manually uniquify titles
and revisions in our title/revision listing end points.