Cleanup & If-Unmodified-Since support
This patch cleans up the code paths in the parsoid module further by combining
the revision fetching in getFormat with the one needed for the equality check
as introduced in the previous patch. This also makes it easier to add
if-unmodified-since support. On precondition failure, status 412 is returned
and no parsing or other content is returned.

Bug: https://phabricator.wikimedia.org/T93777