Use the PHP API for revision and title listings
We ran into issues with Cassandra's native secondary indexes. Large (>64k
IIRC) non-index column values caused inserts to fail with an index on one of
the partition keys. We only used them for fairly hacky listings anyway, so
decided to drop them in restbase-mod-table-cassandra recently.

This patch switches the existing listing to use the PHP API instead. This
gives us time to implement this more cleanly in both the interface & the
backend while still allowing users to discover titles for testing & providing
a glimpse of the direction we are heading in.

Bug: https://phabricator.wikimedia.org/T94148