Pass previous / current revision's content to Parsoid
Parsoid implements several optimizations based on the reuse of existing HTML
content. For example, it reuses template expansions when only images need to
be updated, or vice versa. It also user a previous revision's template
expansions in the rendering of a successor revision.

This patch implements passing this information from RESTBase to Parsoid in a
JSON POST request.

Bug: https://phabricator.wikimedia.org/T92972