Add test config & use it for tests
So far we have tested with our example config, which made it difficult to test
less standard end points and required some ugly request rewriting.

This patch starts to address that by creating a separate config.test.yaml,
which only defines the test domain and pulls in a new test spec for public
test-only entry points. This test spec is separate from the MediaWiki content
API spec, and only adds the test entry points we need.

As a result, we can remove the ugly request rewriting in the restbase HTTP
handler.