Provide full version in etag headers
Clients like the Content Translation service need a generic way to figure out
the revision they received when using end points that aren't revision-based.
Parsoid provided a private header (x-content-revision) for this so far, but in
RB it is preferable to use standard headers instead. In HTTP, revision IDs are
normally provided in the ETag header. We already passed the timeuuid in the
ETag header, but left out the revision so far.

This patch changes the format of our ETag header to include the revision in
addition to the tid, separated by a slash:

ETag: "123456/c4e494da-ee8f-11e4-83a1-8b80de1cde5f"

This provides content-independent version information that's reasonably simple
to parse. Quotes are required by the ETag spec (to do with the w/ prefix for
weak identifiers), sorry about that.

Bug: T97393