Urgent fix: Allow JSON content in POST requests
After parsing a POST request's body, we attempt to parse its contents if
the `application/json` content type is present. However, the code tries
to parse `req.body`, which is undefined, since `rbUtil.parsePOST()`
returns the body, but does not set the `req.body` field. This commit
fixes the issue by correctly using the return value