Added endpoint to translate sections to wikitext (T94890).
A new endpoint /transform/sections/to/wikitext/{title}/{revision}
was added. It takes the original revision from restbase, replaces
sections according to JSON, provided in a sections parameter, and
posts it to parsoid for html->wikitext translation.

A set of tests was added verifying the sections could be replaced
and added, and verifying errors on non-existent sections and invalid
JSON.

Bug: https://phabricator.wikimedia.org/T94890