Added cookie forwarding (T88016)
Cookies set on original request are forwarded to parsoid and
mediawiki endpoints to let user auth and read content.

Two tests added, verifying that cookies are actually forwarded -
one for parsoid and one for mediawiki api.

This solves only a part of the problem, after this fix user can
read content from private wiki, but we still do not have a mechanism
to forbid read of content already stored in restbase.

Bug: https://phabricator.wikimedia.org/T88016