Pass through scrubWikitext to Parsoid
Parsoid added an API flag to fix up HTML in ways that results in cleaner
wikitext after conversion. While it's not clear that this is going to be the
longer-term API, we agreed to pass this through as-is without documenting it
to allow VE and other clients to take advantage of the feature quickly.

We can then discuss the longer-term API design in
https://phabricator.wikimedia.org/T106909.