Hook up Mathoid in RESTBase
This commit adds the public API end-points and the back-end module
definition for proxying calls to Mathoid - the math rendering service.

Two public routes are added:
- `/media/math/{format}/{formula}`
- `/media/math/{format}/{formula}/{type}`

The first renders a formula given in TeX format into the specified
format, while the latter allows a client to specify the input type as
well.

Bug: [T102030](https://phabricator.wikimedia.org/T102030)