Don't perform expensive no-cache instance tests
While doing some latency testing I noticed that the title Foobar was really
slow to load (500ms). I eventually traced this down to many thousands of
revision entries (possibly hundreds of thousands), which are caused by x-ample
tests forcing revision refreshes with no-cache.

Looking at the config, I found another x-ample that continuously re-renders
Foobar's HTML. This only avoided filling up revision storage as the content of
Foobar does not change much between renders, and we have implemented an
optimization that avoids storing unchanged revisions.

In both cases, there is really not much of a point in testing re-renders
separately for each instance. We have plenty of data on backend connections
that we can set alerts on. Having RESTBase alerts go off when the Action API
is in trouble is confusing, which is the last thing we need when things aren't
working so well.