Quote the profile in content-type headers
According to RFC 7231 section 3.1.1, parameter values are either
tokens or quoted strings.  The forward-slash character ("/") is not
allowed in a token (defined in RFC 7230, section 3.2.6).  We need to
make the parameter value a quoted string.

This parsing restriction is actually enforced by the npm
`content-type` module (version 1.0.1) which is used by Express 4.x to
manipulate content type headers.  The unmodified content-type causes
an exception to be thrown.

Change-Id: I6596d4b0c79cbbe435934290ff15fafc8179217a