Require a tid iff a revision is supplied in html2wt
Parsoid uses both the original HTML and data-parsoid in the html2wt
conversion. Each entry in data-parsoid is associated with a DOM node via a
unique ID. These IDs can change randomly between renders, which means that it
is critical for correctness to supply both original HTML and data-parsoid from
the exact same render. This means that we need to require a tid to be supplied
iff the revision is specified.

This patch adds a simple check for the tid, and throws a 400 if it is missing.

Bug: T111491