Templating: Allow URIs to have absolute templates
Many use cases require us to be able to make arbitrary variables part of
the next request's URI. One such instance is the usage of the POST
request's hash in order to determine whether the response to that
request has been previously stored, making the data's hash string the
perfect candidate as the primary key. This commit adds that flexibility
without incurring additional performance penalty for templates not
needing the feature.