Optimized request templating
This PR optimizes request templating further. The templating micro-benchmark
is now about 2-3 times faster, with a further 3x speedup to be had if we
hoisted the try/catch one level.

However, profiles don't really show this code (before or after) as that much
of a bottleneck.  There are some gains to be made here for RESTBase as a
whole, but they will primarily benefit otherwise cheap entry points, where
request rewriting is a bigger part of the overall request cost. For most HTML
requests, much more time is spent in the Cassandra driver.

To summarize, I wouldn't call this change necessarily super-critical for
overall RESTBase performance, at this point.

- Don't silently ignore when entire sub-trees are undefined. This is still
  supported where needed (via default(), for example), but making this the
  default strikes me as dangerous.
- More aggressively expand all simple templates inline by rewriting local
  references to the path prefix. This isn't perfect yet, but does work rather
  well.
- Rename Template.eval to Template.expand; `eval` is not popular with jshint.

- More robust scope rewriting (see inline notes)
- Possibly, reduce the try/catch tax by moving that up to the request wrapper,
  similar to the technique described by the bluebird authors:
  https://github.com/petkaantonov/bluebird/wiki/Optimization-killers#2-unsupported-syntax)
- Reconsider if we really want to default to `method: spec.method ||
  request.method || 'get'`, or just `method: spec.method || 'get'`. This
  strikes me as a little bit too much magic, and it's good to call out
  potentially state-changing requests explicitly.
- Remove commented-out inline microbenchmarking code.