Parsoid mod: Add warn log lines for tranfroms to wikitext
Somewhere in the VE <> VRS <> RESTBase <> Parsoid pipeline something
breaks and 400 errors are returned to VE. This is a temporary patch for
RESTBase so that it emits warn-level log lines when such a situation
occurs.

Note: this patch only lets us get around the production logger
warn-level setting because of which the emitted errors do not make it
through to logstash. That means that RESTBase will in reality emit two
log lines for each of the errors targeted here. Hence, as soon as we
figure out what needs to be fixed this commit should be reverted.