Drop heapdump build dependency
We already add the heapdump install in the docker build script, so we don't
necessarily need to list it as a dev dependency. This way, most users get a
simplified install without any binary modules, while we still have heapdump
available to debug things in production.

This change also avoids spending time on temporary build failures in travis,
which might be related to old libraries or compilers.