MobileApps: Set up RESTBase storage
Thus far, requests for the Mobile Content Service wre only proxied
through RESTBase. This commit sets up RESTBase to store the response
from the back-end service and serve it on subsequent requests.

First, the current revision ID for a page is retrieved from the storage.
Then, the storage is checked for the content rendered by the Mobile
Content Service for the returned revision ID and time UUID. If data for
that combination exists, it is returned. Otherwise, a request is placed
to the Mobile Content Service to render the page and the result is
stored in RESTBase and returned to the client. This principle is applied
to all of the routes exposed by RESTBase for the Mobile Content Service.