MobileApps: Do not check storage for no-cache requests
This commit introduces two additional routes:
- /handling/content/no-cache/{route}/{title}
- /handling/content/{cache-default}/{route}/{title}

After retrieving the revision info, the system branches on one of them
depending on the contents of the `Cache-Control` header. If it's set to
`no-cache`, the former route is picked, which asks the back-end service
to render the content and stores it. Otherwise, the latter route is
used, which checks the storage for an existing render first.