Save space by flattening the per article table
We found that we're using up a lot more disk space than we thought, so
we're going to flatten the per article table.  This change needs to be
deployed so we can create the table on the AQS cluster and start loading
it.  Once it's loading up properly, we can finish the change by doing
all the TODOs in a separate Pull Request.

Please merge and update the gerrit RESTBase mirror and we'll do the rest
of the deploy.