Back-end services: Strip content-length and content-encoding headers
when storing and returning results from back-end services, such as
Mathoid and MobileApps, we store and return all of the headers. Alas,
some them pertain only to the request made by RESTBase, not the one made
by the client. Such headers include `content-encoding` and
`content-length`, so remove them from the response before storing and
returning them.