Router: Add support for templated sub-specifications
Our configuration file has become too long. Moreover, when something
changes in it, we have to be careful to propagate them (correctly) into
labs, staging, production and MW-Vagrant, making the (manual) process
time-consuming and error-prone. This PR tries to minimise config change
impacts by adding parts of the private routes (all usually contained in
the config file) to specs/sys/v1. Concretely, Mathoid's and MobileApps'
/sys/ routes have been moved there.

RESTBase's private routes usually connect it with a back-end service,
whose IP changes depending on the environment it is being run in. Hence,
this PR also introduces the Swig templating engine which is used to
compile and render the subspecs found under the specs/ directory
hierarchy. For each template, the complete service configuration is
available (more precisely, services[0].conf from the service-runner
config is used).