Clean up documentation and small format fix
We would love this to go out with the RESTBase deploy on Monday, since
we're planning on announcing the API then.  I'm a lot happier with the
new documentation and the only code change was to go from YYYYMMDDHH to
YYYYMMDD for the per-article endpoint.  This was needed since we dropped
hourly resolution.  I updated the tests to reflect the change.