Modular specs
This patch adds support for loading specs as a special kind of module. Similar
to @mobrovac's PR #413, this makes our config more modular and less verbose.
As with other modules, spec modules can be parametrized with options.

Example config stanza:

```yaml
    /{module:graphoid}:
      x-modules:
        - path: graphoid_test
          type: spec
          options:
            host: http://graphoid.wikimedia.org
            header_whitelist:
              - content-type
              - x-foobar
```

The `options` specified for the module can be referenced in request templates
under `$$.conf`. Example:

```yaml
uri: '{+$$.conf.host}/{domain}/...'
headers: '{$$.filter($.request.headers, $$.conf.header_whitelist)}'
```
Note that we are passing an array to `$$.filter`, which works as we are
passing around references.