Don't add a fixed prefix to relative module paths
Previously, we would magically prefix `./mods/` to relative module paths. Now
that we are using modules a bit more broadly & are moving out the
swagger-service functionality, it makes sense to do away with this magic &
simply support any path relative to the root directory.

We will actually need to refine this further to be relative to the
service-runner base directory, but that can be done in a separate step.