Mathoid: Use POST only for formula check, GET for renders
The new textvcinfo endpoint in Mathoid is able to check TeX formulae for
correctness and normalised them. The endpoint is rather cheap and quick,
so call only this one for POST requests. The normalised response is
stored into post storage and hashed. As usual, the hash is returned in
the `x-resource-location` header, which can then be used to issue GET
requests to `/media/math/render/{format}/{hash}` to obtain an actual
render.