Make mathoid responses cacheable
Clients will directly hit restbase for SVG renders of mathematical formulas.
The URL contains a hash of the original formula, and the render does not
change very often. For this reason, we can allow Varnish to cache these SVGs
for extended periods of time.

This patch adds a cache-control header to all responses, and makes the actual
value configurable as a module parameter.