Define /sys/ as api roots to scope metrics
API roots define scopes for both the swagger spec and metrics. As we want to
aggregate metrics at the api level (/sys and /v1), we need to define /sys as
an API root as well.

The current logic requires both swagger and info.x-is-api-root attributes to
establish new API spec roots. It feels like we might be able to improve on
this, but the best I can come up with is moving x-is-api-root to the spec
level, and perhaps renaming it. Other ideas?

In any case, we can now tweak this later without affecting third party user's
configs in the normal case.