Perform background updates after saving Parsoid data
Some background updates request Parsoid data we are just updating. Starting
those updates in parallel without waiting for this data to be saved runs the
risk of triggering a second generation of the same data.

This patch changes the processing order so that the Parsoid data is first
fully saved before starting any background updates.