Support streaming responses
Use cases like chunked storage and proxied requests can be sped up by
streaming the response as it comes in.

This patch adds support for returning a `Readable` stream from handlers, and
is targeted at chunked storage buckets. In the longer term, we probably want
to switch to the richer
[ReadableStream](https://streams.spec.whatwg.org/#rs-model) abstraction, but
node implementations for that are still lacking, and this gets us started with
Node's native streams.

Task: https://phabricator.wikimedia.org/T124696