Set up prefixed metric trees
We currently mix metrics for internal and external requests with update
requests. As the performance characteristics of especially update requests
differ significantly, this makes it hard to establish actual performance for
the most important use cases.

This patch uses new child metric producer support in service-runner 1.0 to
separate metrics of primary and derived requests into three trees:

1) external requests,
2) internal update requests, and finally
3) internal service requests.

This is done by setting up two prefixed child metric producers, and passing
those to the handler to use for all metrics reporting when internal / update
requests are identified.