Mathoid: store texvcinfo check results
Mathoid's /texvcinfo endpoint check the correctness of a formula and
normalises it. Because of the normalisation, we weren't saving its
result, since different formulae can have different TeX representations
but the same render. Unfortunately, that implied that Mathoid had to be
called on each request for a formula check. This commit introduces the
mathoid.js sys module which de-duplicates such cases. First, storage is
checked for the texvcinfo result of the provided formula. If one cannot
be found, an indirection record is looked up in a special, hash-mapping
table. In case an indirection exists, that hash is used to complete the
request (the texvcinfo table is queried again). Only if all of the above
fails, Mathoid is called.