Mathoid: Enable PNG support
Mathoid sends the PNG in binary form to the caller. However, when
wrapping the PNG response body in a JSON for its `/complete` endpoint,
it gets serialised to a plain `{type: 'Buffer', data: [...]}` object. To
circumvent that and be able to still make only one request for all of
the renders, requesting the renders from Mathoid and storing them in
separate buckets has been moved to the mathoid module, where the PNG
render gets converted back into a Buffer and stored (and served)
properly.