Option for the bucket type backing the parsoid module
This patch makes the bucket type used to back the parsoid module configurable.
The default type is changed to key_rev_value, which means that this
effectively disables the key_rev_latest_value change for initial deployment.

Once we are satisfied with test results, we can enable key_rev_latest_value
with a config change.