Parsoid: Always use key_rev_value for stashed content
PR #558 introduced the possibility to specify which bucket type to use
for regular Parsoid content. However, stashed content always needs to
use the `key_rev_value` bucket type, given that there is no notion of
'latest' there. As both code paths use `getBucketURI()` to determine the
URI of the underlying table to use, fix the value of the bucket type
when requesting stashed content to `key_rev_value`.