By default, preserve all parameters in relative redirects
Relative redirects are used for different use cases, some of which
require the preservation of all parameters (ex: title normalization), and
others do not. Before eec5180, we defaulted to full preservation, and have
since switched to no preservation. As a result, normalized redirects for
anything but wikitext redirects are currently not fully preserving all
parameters.

This patch makes the path stripping optional with a boolean option. To
generally reduce the number of parameters, an `options` object is introduced.