Make tests more reliable
We have recently seen a lot of spurious travis failures, which consumed a lot
of time for re-running tests & checking whether there were actual failures or
not.

Most of these spurious failures were related to local / remote request
classification, which is based on execution logs. The same tests are
consistently passing locally, so it seems likely that we are seeing some kind
of race condition, which is more likely to trigger when tests are executed
slowly in travis.

This patch changes these tests to check for equality of tids instead. While this
does not directly check for presence of local / remote requests, it does
indirectly test that the content is saved & then served from storage in
subsequent requests.

Another class of failures came from the event production test, which only left
100ms for async request processing, before testing that the event had indeed
been processed. This was not always enough in the travis environment, so this
patch bumps up the timeout to 2000ms.