Start testing rate limiting on transform end points
Since March, we have a basic rate limit filter in the hyperswitch repository.
This is currently only logging when rates are exceeded, so that we can start
by verifying the basic operation without a risk of aborting requests
unnecessarily.

This patch enables this filter with conservative limits on the un-cached
`transform` end points. By deploying this change (and a puppet change to
configure the rate limiter), we can move forward by testing rate limiting in
staging & production.