Limit the maximum rate of requests per IP
Configure the metrics endpoints to filter out requests that happen more
than X times per second from the same IP to the same endpoint.  The
exact number is not decided yet but I'm tentatively setting it at 100.

Right now, the filter is just logging if the limit is exceeded so we can
get a feel for how often this happens.  We may decide to make it a real
filter as part of this PR or later.

https://phabricator.wikimedia.org/T135240
Bug: T135240