Skip sending resource_change events if so wanted (#630)
* Parsoid mod: Skip sending resource_change events if so wanted

RESTBase automatically sends resource_change events when the HTML of a
page has changed, which allows us to track dependant updates. However,
there are circumstances a user might not want them to be sent. There are
two cases where that is true:
- third-party support: these events make sense only when third-party
  users users have the complete EventBus system installed (HTTP Proxy +
  Kafka + ChangeProp), which make the events virtually useless for small
  installs; and
- staging: staging hosts are part of production, which means they emit
  events to the production Kafka Cluster, and is undesirable when, e.g.,
  dumps are running there.

This patch adds the `skip_updates` configuration stanza that bypasses
the Parsoid module's sending of resource_change events when set to
`true` (the default value is `false`).

* Events Mod: Obey the skip_updates config stanza

* Do not use skip_updates in Parsoid mod, only in Events