Add feed and random public endpoints; release v0.14.2 (#636)
* Add feed/featured and page/random/{format} endpoints

`/feed/featured` is the aggregate endpoint used by the next version of
the native Adroid app. The endpoint forwards the request to the Mobile
Content Service and returns its response. There is no storage for it at
the moment; it should be added in a second iteration (once the
endpoint's output format stabilises) and it should be modelled after
mobile-sections since it returns the aggregate response for multiple
smaller endpoints (they will be added once we start storing this
endpoint's response). Varnish is instructed to cache the response for
one minute.

`/page/random/{format}` gets a radnom page title from MCS'
`page/random/title` endpoint and returns a redirect to the client based
on the chosen format. Valid formats are title, html, summary, related,
mobile-sections and mobile-sections lead. Since different random titles
should be returned as often as possible, the redirect is cached for only
2 seconds.

Bug: (T136960)[https://phabricator.wikimedia.org/T136960]

* Monitoring spec: Test handling improvements

- if the example request does not specify the domain, assume it's the
  domain we are currently testing for
- don't set a field value to '' if it is false-y, but only if it's null
  or undefined
- if the example's response contains an array with only one element,
  assume we want to check each item of the returned array against it

* Release v0.14.2