Feeds: Split the aggregated feed request and cache it (#662)
Instead of requesting the whole response of the aggregated feed endpoint
in MCS, it is much faster to request its individual pieces and then
assemble them in RESTBase. The process happens in parallel, so different
MCS workers get to work on a part of the response.

Moreover, in order to decrease the amount of requests to MCS, store the
response using the `ttl` revision retention policy. Currently, we begin
with the conservative ttl of 10 minutes, but it is likely we can
increase it in the future. The random component needs to change more
frequently than that, so whenever we have a cache hit, we re-request a
random title from MCS and incorporate it in the new response. We also
tell Varnish to cache the response for 30 seconds. This means that the
number of requests will be:
- 1 request for the random title every 30 seconds
- 5 requests for the various response components every 10 minutes